# OpenAPI/payment_system-openapi_user.yaml
openapi: "3.1.0"
info:
  title: "payment_system API"
  description: "payment_system API"
  version: "1.0.0"
servers:
  - url: "http://localhost:8080"
paths:
  /api/v1/user:
    post:
      summary: "POST api/v1/user"
      operationId: "registerUser"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegistrationRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/User"
  /api/v1/user/{userId}:
    get:
      summary: "GET api/v1/user/{userId}"
      operationId: "getUserByUserId"
      parameters:
        - name: "userId"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/User"
  /api/v1/users/{userId}/payments:
    get:
      summary: "GET api/v1/users/{userId}/payments"
      operationId: "getPaymentsforUserId"
      parameters:
        - name: "userId"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
        - name: "page"
          in: "query"
          required: false
          schema:
            type: "integer"
            format: "int32"
            default: "0"
        - name: "pageSize"
          in: "query"
          required: false
          schema:
            type: "integer"
            format: "int32"
            default: "20"
        - name: "startDate"
          in: "query"
          required: false
          schema:
            type: "string"
            format: "date-time"
        - name: "endDate"
          in: "query"
          required: false
          schema:
            type: "string"
            format: "date-time"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaymentDTO"
components:
  schemas:
    BigDecimal:
      type: "number"
    UserRegistrationRequest:
      type: "object"
      properties:
        initialBalance:
          $ref: "#/components/schemas/BigDecimal"
    User:
      type: "object"
      properties:
        userId:
          type: "string"
          format: "uuid"
        initialBalance:
          $ref: "#/components/schemas/BigDecimal"
        balance:
          $ref: "#/components/schemas/BigDecimal"
        createdAt:
          type: "string"
          format: "date-time"
    PaymentDTO:
      type: "object"
      properties:
        paymentId:
          type: "string"
          format: "uuid"
        senderId:
          type: "string"
          format: "uuid"
        receiverId:
          type: "string"
          format: "uuid"
        amount:
          $ref: "#/components/schemas/BigDecimal"
        timestamp:
          type: "string"
          format: "date-time"